fix paymentInfo + refactor calls
