+++ src/pages/gift/generate/ConfirmGift.js
@@ -21,9 +21,10 @@ export default function ConfirmGift ({ account, giftInfo, generateGiftHandler, g
         const address = account?.address;
         if (address && api) {
           const chainAmount = utils.toChainUnit(amount, chainInfo?.decimals);
-          const info = await api.tx.balances
-            .transfer(address, chainAmount || 0)
-            .paymentInfo(address);
+          const transferTx = api.tx.balances.transfer(address, chainAmount || 0);
+          const remarkTx = api.tx.system.remarkWithEvent('gift:send');
+          const txs = [transferTx, remarkTx];
+          const info = await api.tx.utility.batchAll(txs).paymentInfo(address);
 
           let estimatedFee = utils.calcFeeAdjustments(info?.partialFee);
           estimatedFee = estimatedFee.muln(giftFeeMultiplier);
